Monitoring instructions update (#166)
* remove the binary cromwell_metadata_bq;
everyone should build their local version

* new way to query cost (now directly through an R api);
old way Rscript moved to legacy

* update get_metadata_and_runtime_metrics.md, and
rename it to collecting_metadata_and_runtime_metrics.md

deleted legacy SQL instructions, and
a new doc with instructions on how to query into
all three tables of the BQ dataset

* trivial files move
